+++ ts/beamcoder.ts
@@ -1,0 +1,5 @@
+import bindings from 'bindings';
+import { BeamcoderType } from './types/BeamcoderType';
+
+const beamcoder = bindings('beamcoder') as BeamcoderType;
+export default beamcoder;
